s3 Bug1

+++ src/route/uploadingFiles/index.js
@@ -1,64 +1,64 @@
 import fs from "fs";
 import AWS from "aws-sdk";
 import s3Zip from "s3-zip";
+import path from "path";
+const join = path.join;
 import ContractModel from "../../model/contractModel.js";
 const uploadFile = async (req, res) => {
   console.log("uploadFile", req.files);
-  const { id } = req.body;
+  const { id } = req.query;
+  console.log("id", id);
+
   if (req.files.length > 0) {
     try {
-      const contractExist = await ContractModel.findOne({
-        id: id,
-      });
-      if (contractExist) {
-        res.send({ ok: false, message: "Contract ID already exist" });
-      } else {
-        if (req.files) {
-          const s3 = new AWS.S3({
-            accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-            secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-          });
-          console.log("files", req.files.length);
-          for (let i = 0; i < req.files.length; i++) {
-            s3.upload(
-              {
-                Bucket: process.env.AWS_BUCKET_NAME,
-                Body: fs.createReadStream(req.files[i].path),
-                Key: `${id}/${req.files[i].originalname}`,
-              },
-              (err, data) => {
-                if (err) {
-                  console.log("error", err);
-                } else {
-                  console.log("data", data);
-                  req.files.forEach((file) => {
-                    fs.unlink(file.path, (err) => {
-                      console.log("error", err);
-                    });
+      if (req.files && req.files.length > 0) {
+        const s3 = new AWS.S3({
+          accessKeyId: process.env.AWS_ACCESS_KEY_ID,
+          secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
+        });
+        console.log("files", req.files.length);
+        for (let i = 0; i < req.files.length; i++) {
+          s3.upload(
+            {
+              Bucket: process.env.AWS_BUCKET_NAME,
+              Body: fs.createReadStream(req.files[i].path),
+              Key: `${id}/${req.files[i].originalname}`,
+            },
+            (err, data) => {
+              if (err) {
+                console.log("error", err);
+              } else {
+                console.log("data", data);
+                req.files.forEach((file) => {
+                  fs.unlink(file.path, (err) => {
+                    console.log("error", err);
                   });
-                }
+                });
               }
-            );
-          }
-        } else {
-          res.send({ ok: false, message: "No file to upload" });
+            }
+          );
         }
+      } else {
+        res.send({ ok: false, message: "No file to upload" });
       }
     } catch (e) {
       console.log("Error", e);
     }
-    res.send({ ok: true, files: req.files });
+    res.send({ ok: true, message: "File is uploaded" });
   } else {
     res.send({ ok: false, message: "No files" });
   }
 };
 const checkContractExist = async (req, res, next) => {
-  const { id } = req.body;
+  const { id } = req.query;
+  console.log("id", id);
+
   try {
     const docExist = await ContractModel.findOne({ id: id });
     if (docExist) {
       res.status(400).send({ ok: false, message: "Document already exist" });
     } else {
+      console.log("docExist", docExist);
       next();
     }
   } catch (error) {
@@ -74,22 +74,46 @@ const getFiles = async (req, res) => {
     const s3 = new AWS.S3({
       accessKeyId: process.env.AWS_ACCESS_KEY_ID,
       secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-      signatureVersion: "v4",
-      s3ForcePathStyle: true,
     });
-    s3Zip
-      .setFormat("tar")
-      .archive(
-        {
-          s3: s3,
-          region: process.env.LOCATION,
-          bucket: process.env.AWS_BUCKET_NAME,
-          debug: true,
-        },
-        folder,
-        fileNames
-      )
-      .pipe(res);
+    const files = s3
+      .listObjects({
+        Bucket: process.env.AWS_BUCKET_NAME,
+        Prefix: folder,
+      })
+      .createReadStream();
+    function zip(files) {
+      console.log(files);
+      // const output = fs.createWriteStream(join(__dirname, "use-s3-zip.zip"));
+      s3Zip
+        .setFormat("tar")
+        .archive(
+          {
+            region: process.env.LOCATION,
+            bucket: process.env.AWS_BUCKET_NAME,
+            preserveFolderStructure: true,
+          },
+          folder,
+          files
+        )
+        .pipe(res);
+    }
+    zip(fileNames);
+    // const file = s3Zip
+    //   .archive(
+    //     {
+    //       s3: s3,
+    //       region: process.env.LOCATION,
+    //       bucket: process.env.AWS_BUCKET_NAME,
+    //       debug: true,
+    //     },
+    //     folder,
+    //     fileNames
+    //   )
+    //   .pipe(res)
+    //   .on("finish", () => {
+    //     console.log("finish");
+    //   });
+    // console.log("file Here", file);
   }
 };
